Created a new uploadForm page and connected it to Supabase bucket.
New page allows users to define the item name, description and price. At this stage the submission only uploads the file to a bucket. In future we would want this to also create a row in the items table with a link to the supabase bucket. The rows will then be called when a user renders the gallery.