Refactored fetch collections behaviour and created loading screen.
Previously, fetch collections only ran when attempting to create collections. Caused a pause when user tried to create a collection since we needed to reach out to db. Moved fetch collections into useEffect hook that runs on mount, and on successful creation of collections. Also included a loading screen to hold user's attention while we fetch this data on mount.