Building individual collection page and connecting with database.
Have now extended the individual collection page so that it retrieves the images from the database based on the uuid of the collection. Included ability to add new images to the collection. Removed pages that were doubled up including uploadForm and backup collection page. Need to work on: how to close the upload popover when an upload is successful + making the add image button sticky + introduce a delete button with confirmation modal.